chore(config): split monolithic config into domain schemas with Zod (#4232)
